# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "tcmrsv",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/authorize": {
      "post": {
        "summary": "認証",
        "deprecated": false,
        "description": "",
        "operationId": "authorize",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "user_id",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Authorization"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/authorize/refresh": {
      "post": {
        "summary": "再認証",
        "deprecated": false,
        "description": "",
        "operationId": "reauthorize",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                },
                "required": [
                  "refresh_token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Authorization"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/rooms": {
      "get": {
        "summary": "練習室の取得",
        "deprecated": false,
        "description": "",
        "operationId": "getRooms",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RoomList"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reservations/create": {
      "post": {
        "summary": "予約の作成",
        "deprecated": false,
        "description": "",
        "operationId": "createReservation",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "campus_code": {
                    "type": "string",
                    "enum": [
                      "1",
                      "2"
                    ]
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "from_hour": {
                    "type": "integer"
                  },
                  "from_minute": {
                    "type": "integer"
                  },
                  "to_hour": {
                    "type": "integer"
                  },
                  "to_minute": {
                    "type": "integer"
                  },
                  "is_auto_select": {
                    "type": "boolean"
                  },
                  "room_id": {
                    "type": "string"
                  },
                  "booker_name": {
                    "type": "string"
                  },
                  "piano_numbers": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "piano_types": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "grand",
                        "upright",
                        "none"
                      ]
                    }
                  },
                  "floors": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "is_basement": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "campus_code",
                  "date",
                  "from_hour",
                  "from_minute",
                  "to_hour",
                  "to_minute",
                  "is_auto_select"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Reservation"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reservations/{reservation_id}/delete": {
      "delete": {
        "summary": "予約の削除",
        "deprecated": false,
        "description": "",
        "operationId": "deleteReservation",
        "tags": [],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response",
                  "required": [
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reservations/{reservation_id}/update": {
      "put": {
        "summary": "予約の更新",
        "deprecated": false,
        "description": "",
        "operationId": "updateReservation",
        "tags": [],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "campus_code": {
                    "type": "string",
                    "enum": [
                      "1",
                      "2"
                    ]
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "from_hour": {
                    "type": "integer"
                  },
                  "from_minute": {
                    "type": "string"
                  },
                  "to_hour": {
                    "type": "string"
                  },
                  "to_minute": {
                    "type": "string"
                  },
                  "is_auto_select": {
                    "type": "boolean"
                  },
                  "room_id": {
                    "type": "string"
                  },
                  "booker_name": {
                    "type": "string"
                  },
                  "piano_numbers": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "piano_types": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "grand",
                        "upright",
                        "none"
                      ]
                    }
                  },
                  "floors": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "is_basement": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "campus_code",
                  "date",
                  "from_hour",
                  "from_minute",
                  "to_hour",
                  "to_minute",
                  "is_auto_select"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response",
                  "required": [
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reservations/{reservation_id}": {
      "get": {
        "summary": "予約の取得",
        "deprecated": false,
        "description": "",
        "operationId": "getReservation",
        "tags": [],
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Reservation"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reservations/mine": {
      "get": {
        "summary": "自分の予約を取得",
        "deprecated": false,
        "description": "",
        "operationId": "getMyReservations",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReservationList"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "piano_type": {
            "type": "string",
            "enum": [
              "grand",
              "upright",
              "none"
            ]
          },
          "piano_number": {
            "type": "integer"
          },
          "is_classroom": {
            "type": "boolean"
          },
          "is_basement": {
            "type": "boolean"
          },
          "campus_code": {
            "type": "string",
            "enum": [
              "1",
              "2"
            ]
          },
          "floor": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "id",
          "piano_number",
          "piano_type",
          "is_classroom",
          "is_basement",
          "campus_code",
          "floor"
        ]
      },
      "RoomList": {
        "type": "object",
        "properties": {
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Room"
            }
          }
        },
        "required": [
          "rooms"
        ]
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "external_id": {
            "type": "string"
          },
          "campus_code": {
            "type": "string",
            "enum": [
              "1",
              "2"
            ]
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "room_id": {
            "type": "string"
          },
          "from_hour": {
            "type": "integer"
          },
          "from_minute": {
            "type": "integer"
          },
          "to_hour": {
            "type": "integer"
          },
          "to_minute": {
            "type": "integer"
          },
          "booker_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "campus_code",
          "date",
          "room_id",
          "from_hour",
          "from_minute",
          "to_hour",
          "to_minute",
          "created_at"
        ]
      },
      "ReservationList": {
        "type": "object",
        "properties": {
          "reservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            }
          }
        },
        "required": [
          "reservations"
        ]
      },
      "Authorization": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
      "Response": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "code": {
            "type": "integer"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "ok",
          "data",
          "code"
        ]
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}